compose: Split off a containers_storage module
Prep for further work.
