+++ rust/src/compose.rs
@@ -6,7 +6,7 @@ use std::ffi::OsStr;
 use std::fs::File;
 use std::io::{BufWriter, Write};
 use std::os::fd::{AsFd, AsRawFd};
-use std::process::{Command, Stdio};
+use std::process::Command;
 
 use anyhow::{anyhow, Context, Result};
 use camino::{Utf8Path, Utf8PathBuf};
@@ -23,6 +23,7 @@ use ostree_ext::{container as ostree_container, glib};
 use ostree_ext::{oci_spec, ostree};
 
 use crate::cmdutils::CommandRunExt;
+use crate::containers_storage::PodmanMount;
 use crate::cxxrsutil::{CxxResult, FFIGObjectWrapper};
 use crate::isolation::self_command;
 
@@ -187,89 +188,6 @@ pub(crate) struct BuildChunkedOCIOpts {
     output: String,
 }
 
-struct PodmanMount {
-    path: Utf8PathBuf,
-    temp_cid: Option<String>,
-    mounted: bool,
-}
-
-impl PodmanMount {
-    #[context("Unmounting container")]
-    fn _impl_unmount(&mut self) -> Result<()> {
-        if self.mounted {
-            tracing::debug!("unmounting {}", self.path.as_str());
-            self.mounted = false;
-            Command::new("umount")
-                .args(["-l", self.path.as_str()])
-                .stdout(Stdio::null())
-                .run()
-                .context("umount")?;
-            tracing::trace!("umount ok");
-        }
-        if let Some(cid) = self.temp_cid.take() {
-            tracing::debug!("rm container {cid}");
-            Command::new("podman")
-                .args(["rm", cid.as_str()])
-                .stdout(Stdio::null())
-                .run()
-                .context("podman rm")?;
-            tracing::trace!("rm ok");
-        }
-        Ok(())
-    }
-
-    #[context("Mounting continer {container}")]
-    fn _impl_mount(container: &str) -> Result<Utf8PathBuf> {
-        let mut o = Command::new("podman")
-            .args(["mount", container])
-            .run_get_output()?;
-        let mut s = String::new();
-        o.read_to_string(&mut s)?;
-        while s.ends_with('\n') {
-            s.pop();
-        }
-        tracing::debug!("mounted container {container} at {s}");
-        Ok(s.into())
-    }
-
-    #[allow(dead_code)]
-    fn new_for_container(container: &str) -> Result<Self> {
-        let path = Self::_impl_mount(container)?;
-        Ok(Self {
-            path,
-            temp_cid: None,
-            mounted: true,
-        })
-    }
-
-    fn new_for_image(image: &str) -> Result<Self> {
-        let mut o = Command::new("podman")
-            .args(["create", image])
-            .run_get_output()?;
-        let mut s = String::new();
-        o.read_to_string(&mut s)?;
-        let cid = s.trim();
-        let path = Self::_impl_mount(cid)?;
-        tracing::debug!("created container {cid} from {image}");
-        Ok(Self {
-            path,
-            temp_cid: Some(cid.to_owned()),
-            mounted: true,
-        })
-    }
-
-    fn unmount(mut self) -> Result<()> {
-        self._impl_unmount()
-    }
-}
-
-impl Drop for PodmanMount {
-    fn drop(&mut self) {
-        tracing::trace!("In drop, mounted={}", self.mounted);
-        let _ = self._impl_unmount();
-    }
-}
-
 /// Generate a filesystem tree in ostree-container format from an input manifest.
 /// This can then be copied into e.g. a `FROM scratch` container image build.
 #[derive(Debug, Parser)]
@@ -322,7 +240,7 @@ impl BuildChunkedOCIOpts {
         };
         let rootfs = match &rootfs_source {
             FileSource::Rootfs(p) => p.as_path(),
-            FileSource::Podman(mnt) => mnt.path.as_path(),
+            FileSource::Podman(mnt) => mnt.path(),
         };
         let rootfs = Dir::open_ambient_dir(rootfs, cap_std::ambient_authority())
             .with_context(|| format!("Opening {}", rootfs))?;
