+++ rust/src/lib.rs
@@ -963,6 +963,7 @@ pub mod client;
 pub(crate) use client::*;
 pub mod cliwrap;
 pub mod container;
+mod containers_storage;
 pub use cliwrap::*;
 pub(crate) use container::*;
 mod compose;
